host: extract HostPortUser code from credentials and config
